move paths
