--- src/platform/packages/private/kbn-screenshotting-server/src/paths.ts
@@ -1,135 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-
-import path from 'path';
-
-export interface PackageInfo {
-  platform: 'linux' | 'darwin' | 'win32';
-  architecture: 'x64' | 'arm64';
-  archiveFilename: string;
-  archiveChecksum: string;
-  binaryChecksum: string;
-  binaryRelativePath: string;
-  isPreInstalled: boolean;
-  location: 'custom' | 'chromeForTesting';
-}
-
-enum BaseUrl {
-  // A GCS bucket under the Kibana team
-  custom = 'https://storage.googleapis.com/headless_shell',
-  // GCS bucket for headless chrome provided by the chrome team, see
-  // https://github.com/GoogleChromeLabs/chrome-for-testing#json-api-endpoints
-  chromeForTesting = 'https://storage.googleapis.com/chrome-for-testing-public',
-}
-
-interface CustomPackageInfo extends PackageInfo {
-  location: 'custom';
-}
-
-interface ChromeForTestingPackageInfo extends PackageInfo {
-  version: string;
-  location: 'chromeForTesting';
-  archivePath: string;
-}
-
-function isChromeForTestingPackage(p: PackageInfo): p is ChromeForTestingPackageInfo {
-  return p.location === 'chromeForTesting';
-}
-
-export class ChromiumArchivePaths {
-  public readonly packages: Array<CustomPackageInfo | ChromeForTestingPackageInfo> = [
-    {
-      platform: 'darwin',
-      architecture: 'x64',
-      archiveFilename: 'chrome-headless-shell-mac-x64.zip',
-      archiveChecksum: '772b1fec39826e0604d21e515ee27c47bdfb7a5d69e8674d818b0c159c3dfcb2',
-      binaryChecksum: '493f63172a7b76d12c844427bfdc67d391175e42cba99025277ffce31a7587f7',
-      binaryRelativePath: 'chrome-headless-shell-mac-x64/chrome-headless-shell',
-      version: '136.0.7103.49',
-      location: 'chromeForTesting',
-      archivePath: 'mac-x64',
-      isPreInstalled: false,
-    },
-    {
-      platform: 'darwin',
-      architecture: 'arm64',
-      archiveFilename: 'chrome-headless-shell-mac-arm64.zip',
-      archiveChecksum: '30c5215f2b3caa3321de7c0e0de7225b13cbdd970f7c3f68c774be0ca34fc4e3',
-      binaryChecksum: '6ada593813656a62560680f510a34f0937537ce8518cb60dcdd0c3ceace97f50',
-      binaryRelativePath: 'chrome-headless-shell-mac-arm64/chrome-headless-shell',
-      version: '136.0.7103.49',
-      location: 'chromeForTesting',
-      archivePath: 'mac-arm64',
-      isPreInstalled: false,
-    },
-    {
-      platform: 'linux',
-      architecture: 'x64',
-      archiveFilename: 'chromium-031848b-locales-linux_x64.zip',
-      archiveChecksum: '2f8ede6c874cbf71f6d64ad5c88b33e0f91bdddb4f5684fa6148702ce85550d7',
-      binaryChecksum: '12ba32eadf7dc1e3bd2c72707dfc100d4c9dd6eaddd5568f407a2ed66156e1d1',
-      binaryRelativePath: 'headless_shell-linux_x64/headless_shell',
-      location: 'custom',
-      isPreInstalled: true,
-    },
-    {
-      platform: 'linux',
-      architecture: 'arm64',
-      archiveFilename: 'chromium-031848b-locales-linux_arm64.zip',
-      archiveChecksum: 'de7571709d3b25cd15c48949bb5d72665b98ed527d73dbffbb96116dd5aeaae9',
-      binaryChecksum: 'c2cee1a7906e0c0905153bcc40ece031be5028d075b88736629280ac80877246',
-      binaryRelativePath: 'headless_shell-linux_arm64/headless_shell',
-      location: 'custom',
-      isPreInstalled: true,
-    },
-    {
-      platform: 'win32',
-      architecture: 'x64',
-      archiveFilename: 'chrome-headless-shell-win64.zip',
-      archiveChecksum: 'a4a82311596166c6148df6d1d1497d9bb5397895c47f67abcde4b368180418bc',
-      binaryChecksum: '574d3b27846bdeaafb73632f1f6aac001dc895165465ae11bcd3fedb35cfa9dd',
-      binaryRelativePath: path.join('chrome-headless-shell-win64', 'chrome-headless-shell.exe'),
-      version: '136.0.7103.49',
-      location: 'chromeForTesting',
-      archivePath: 'win64',
-      isPreInstalled: true,
-    },
-  ];
-
-  // zip files get downloaded to a .chromium directory in the kibana root
-  public readonly archivesPath = path.resolve(__dirname, '../../../../../../.chromium');
-
-  public find(platform: string, architecture: string, packages: PackageInfo[] = this.packages) {
-    return packages.find((p) => p.platform === platform && p.architecture === architecture);
-  }
-
-  public resolvePath(p: PackageInfo) {
-    // adding architecture to the path allows it to download two binaries that have the same name, but are different architecture
-    return path.resolve(this.archivesPath, p.architecture, p.archiveFilename);
-  }
-
-  public getAllArchiveFilenames(): string[] {
-    return this.packages.map((p) => this.resolvePath(p));
-  }
-
-  public getDownloadUrl(p: PackageInfo) {
-    if (isChromeForTestingPackage(p)) {
-      const { chromeForTesting } = BaseUrl;
-      const { archivePath, version, archiveFilename } = p;
-      // returned string matches download value found at the following endpoint;
-      // https://googlechromelabs.github.io/chrome-for-testing/known-good-versions-with-downloads.json
-      return `${chromeForTesting}/${version}/${archivePath}/${archiveFilename}`;
-    }
-
-    return BaseUrl.custom + '/' + p.archiveFilename; // revision is not used for URL if package is a custom build
-  }
-
-  public getBinaryPath(p: PackageInfo, chromiumPath: string) {
-    return path.join(chromiumPath, p.binaryRelativePath);
-  }
-}
